Add files via upload
The dataset chosen was “car lemon dataset.csv” (Figure 1). The data included various features of cars (such as color and make) and whether the cars were determined to be good or bad purchases. The dataset included 34 variables and 72,983 rows or observations. The type of variables included numeric (n = 15) and character (n = 19) variables. Of the numeric variables, 11 were interval, the other two were binary. 
The purpose of the analysis was to identify the bad vehicle purchases, from good vehicle purchases.  Ensemble models are “models of models” based on the concept that large numbers of models are more accurate than any one model (Surowiecki, 2005). 
“Is Bad Buy” was the target binary variable (“0” = not a bad buy, “1” = bad buy).  The target variable was unevenly proportioned and skewed heavily in favor of not bad purchases (n = 64,007; 87.70%). Bad purchases totaled 8,976 of the 72,983 observations (12.30%). The imbalance percentage is 75.4%.